requirements.txt
discord.py
requests
beautifulsoup4
googletrans==3.1.0a0